fix broken build caused by 434:3628e68fda20

+++ shell/SystemClass.cpp
@@ -71,7 +71,7 @@ namespace avmshell
 		#ifdef PERFORMANCE_GETTIMER
 		initialTime = MMgc::GC::getPerformanceCounter();
 		#else
-		initialTime = (double)OSDep::currentTimeMillis();		
+		initialTime = OSDep::currentTimeMillis();		
 		#endif // PERFORMANCE_GETTIMER
 
 	}
